Web scraping to find news and articles & predict if they could go viral
Used BeautifulSoup to scrape three global and indian websites, for news content as well as past 'viral' contents- these formed the labeled training. Then taken a random global news website to scrape for test dataset. Used SVM -> not satisfactory, then used KMeans on 2 clusters and it was interesting what it predicted as 'viral'.